refactor: uses `next-connect` in `/migrations` and `/status`
